+++ backend/src/main/java/pt/ua/deti/ies/SmartHomes/backend/HouseData/CostData.java
@@ -1,0 +1,15 @@
+package pt.ua.deti.ies.SmartHomes.backend.HouseData;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+import java.time.Instant;
+
+@AllArgsConstructor
+@Getter
+class CostData {
+    private Instant time;
+    private long electricity;
+    private long water;
+    private long total;
+}
