+++ backend/src/main/java/pt/ua/deti/ies/SmartHomes/backend/HouseData/HouseDataController.java
@@ -4,8 +4,6 @@ import com.influxdb.client.QueryApi;
 import com.influxdb.query.FluxRecord;
 import com.influxdb.query.FluxTable;
 import lombok.AllArgsConstructor;
-import lombok.Getter;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -14,61 +12,8 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import pt.ua.deti.ies.SmartHomes.backend.Devices.DeviceType;
 
-import java.time.Instant;
 import java.util.ArrayList;
 import java.util.List;
-
-@AllArgsConstructor
-@Getter
-class ElectricityData {
-    private Instant time;
-    private long grid_renewable;
-    private long house_solar;
-    private long house_wind;
-    private long house_grid_exchange;
-    private long house_total;
-    private long house_self_sufficiency;
-    private long house_renewable;
-}
-
-@AllArgsConstructor
-@Getter
-class EnvironmentData {
-    private Instant time;
-    private long self_sufficiency;
-    private long renewable;
-    private long emissions;
-    private List<Long> renewable_forecast_day;
-    private List<Long> renewable_forecast_hour;
-}
-
-@AllArgsConstructor
-@Getter
-class WaterData {
-    private Instant time;
-    private long kitchen;
-    private long bath;
-    private long garden;
-    private long other;
-    private long total;
-    private long forecast_today;
-}
-
-@AllArgsConstructor
-@Getter
-class CostData {
-    private Instant time;
-    private long electricity;
-    private long water;
-    private long total;
-}
-
-@AllArgsConstructor
-@Getter
-class DevicePower {
-    private int id;
-    private long power;
-}
 
 @RestController
 @AllArgsConstructor
