+++ backend/src/main/java/pt/ua/deti/ies/SmartHomes/backend/HouseData/WaterData.java
@@ -1,0 +1,18 @@
+package pt.ua.deti.ies.SmartHomes.backend.HouseData;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+import java.time.Instant;
+
+@AllArgsConstructor
+@Getter
+class WaterData {
+    private Instant time;
+    private long kitchen;
+    private long bath;
+    private long garden;
+    private long other;
+    private long total;
+    private long forecast_today;
+}
